+++ ft_memset.c
@@ -6,11 +6,13 @@
 /*   By: juramos <juramos@student.42madrid.com>     +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/09/18 11:19:23 by juramos           #+#    #+#             */
-/*   Updated: 2023/09/18 11:28:12 by juramos          ###   ########.fr       */
+/*   Updated: 2023/09/18 13:06:29 by juramos          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-void	*ft_memset(void *b, int c, int len)
+#include "libft.h"
+
+void	*ft_memset(void *b, int c, size_t len)
 {
 	void	*ret;
